via 優化

+++ BuildVisingGraph.h
@@ -14,7 +14,7 @@ struct Edge{
 	//建樹只需要u,v,cost
 	size_t u;
 	size_t v;
-	u64 cost;
+	u32 cost;
 	Edge(size_t ori_u,size_t ori_v,u8 type):ori_u(ori_u),ori_v(ori_v),type(type){}
 };
